requirements.txt
# requirements.txt

anthropic
backoff
bs4
flask
httpx
nltk
openai>=1.0.0
opencv-python
pandas
playwright
playwright_stealth
tiktoken
google-generativeai
google-ai-generativelanguage
langchain-google-genai
Flask-Session
chromadb
sentence-transformers
wikipedia-api
tqdm
langchain
huggingface_hub
networkx
fastcoref
datasets
scikit-learn
numpy
rdflib
requests-file
rdflib-jsonld
pyld
celery
redis
PyYAML
python-dotenv
beautifulsoup4
requests
trafilatura>=1.6.0